personal-database-assistant/api_bridge/requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
pyyaml==6.0
python-dotenv==1.0.0
cryptography==41.0.5 